requirements.txt
# SmartAttend – Python dependencies
# Tested on CPython 3.12 (Windows 10/11 + PostgreSQL 14+).
#
# Install with:
#     python -m pip install -r requirements.txt
#
# NOTE: face_recognition declares `dlib` as its dependency. On Windows we use
# the precompiled `dlib-bin` wheel instead (no C++ compiler required), so
# `face_recognition` is installed with `--no-deps` further down. Pip will
# still print a "missing dlib" warning – it is safe to ignore because the
# `dlib` Python module is provided by `dlib-bin`.

# --- Backend (FastAPI + SQLAlchemy + PostgreSQL) ---------------------------
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pydantic>=2.5,<3.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# --- Backend yardımcıları --------------------------------------------------
# pandas: opsiyonel rapor/CSV indirme yardımcıları için kullanılabilir.
# requests: bazı seed/sınama scriptleri (örn. local smoke-tests) için.
pandas>=2.1,<3.0
requests>=2.31,<3.0

# Not: Frontend artık React/Vite (Node.js) ile geliştirilmiştir. Eski
# Streamlit prototipi `frontend-streamlit/` klasöründe yedeklidir; orayı
# çalıştırmak için aşağıdaki paketler gerekir (varsayılan kurulumda
# YÜKLENMEZ). İhtiyaç hâlinde elle yükleyin:
#     pip install "streamlit>=1.31,<2.0" "streamlit-autorefresh>=1.0.1"

# --- AI / Computer Vision -------------------------------------------------
# numpy<2 because face_recognition / dlib still rely on legacy NumPy ABI.
numpy>=1.26,<2.0
opencv-python>=4.9,<5.0
# Precompiled dlib wheel (no C++ compiler needed on Windows).
dlib-bin>=19.24.6

# `face_recognition_models` ships large *.dat files using the legacy
# `pkg_resources` import, which was removed in setuptools 81. Keep
# setuptools < 80 so the import keeps working.
setuptools<80
face_recognition_models>=0.3.0

# `face_recognition` itself – installed with --no-deps via post-install
# command (see README). To install everything in one go you can run:
#     pip install -r requirements.txt
#     pip install --no-deps face_recognition
face_recognition>=1.3.0
